Refactor[submodule] #1: Initialize controller & validator of User
